Update week5_tokenization.ipynb
In the SplitMergeFromLogitsTokenizer: inside the code block on code line number 4, the result of tokenizer.tokenize(strings, labels) is not assigned to any variable. So when we use print(decode_utf8_tensor(tokens)), it takes the variable 'tokens' from the previous code cell and not this one. So it should be tokens = tokenizer.tokenize(strings, labels)